pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.43.0",
    "bcrypt>=4.0.1",
    "cryptography>=41.0.0",
    "flask>=3.0.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "streamlit-authenticator>=0.3.3",
    "ta>=0.11.0",
    "yfinance>=0.2.66",
    "anthropic>=0.30.0",
]
